+++ project-model/src/main/java/com/walmartlabs/concord/project/Constants.java
@@ -115,11 +115,6 @@ public final class Constants {
         public static final String REQUEST_DATA_FILE_NAME = "_main.json";
 
         /**
-         * Default request data. File that will be merged with {@link #REQUEST_DEFAULTS_FILE_NAME}'s JSON.
-         */
-        public static final String REQUEST_DEFAULTS_FILE_NAME = "_defaults.json";
-
-        /**
          * Directory, containing job "attachments": reports, stats, etc.
          */
         public static final String JOB_ATTACHMENTS_DIR_NAME = "_attachments";