+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/model/YamlDockerStep.java
@@ -2,16 +2,20 @@ package com.walmartlabs.concord.project.yaml.model;
 
 import com.fasterxml.jackson.core.JsonLocation;
 
+import java.util.Map;
+
 public class YamlDockerStep extends YamlStep {
 
     private final String image;
     private final String cmd;
+    private final Map<String, Object> env;
 
-    public YamlDockerStep(JsonLocation location, String image, String cmd) {
+    public YamlDockerStep(JsonLocation location, String image, String cmd, Map<String, Object> env) {
         super(location);
 
         this.image = image;
         this.cmd = cmd;
+        this.env = env;
     }
 
     public String getImage() {
@@ -20,5 +24,9 @@ public class YamlDockerStep extends YamlStep {
 
     public String getCmd() {
         return cmd;
+    }
+
+    public Map<String, Object> getEnv() {
+        return env;
     }
 }